video: tegra: host: undo boot retry WAR
As part of Bug 1588951 a repeated boot retry WAR was added to mitigate
a rare VIC boot failure. However, there has not been repeated failures
seen on Mobile Sanity of that sort ever since the WAR was
introduced. Furthermore, the WAR was never concluded to be helpful as
the failure was seen one time only.

Clean up the WAR and restore the code back to the simple state it used
to be. Also get rid of handling poweron_toggle_slcg as the slcg
handling has fully moved to power gating driver on recent kernel
versions.

Bug 2485656

Change-Id: Iebeb10b2380d8b38c36956a39c4144b909519610
Signed-off-by: Timo Alho <talho@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2005656
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>