crypto: tegra: use enum for algo.
algo is unsigned char* variable in tegra_sha_req_shash which is a
pointer in userspace and since there are no checks done on this string.
A simple modification of this string can cause a system crash.

Example: passing algo as NULL or a not NULL terminated string.

Avoid this by representing algo by enum in tegra_sha_req_shash.
This will not only avoid most of the checks that we had to do on a string
but it will also improve the execution time.

Example:
	for (i = 0; i < 5; i++) {
		if (!strcmp(sha_req_shash->algo, sha_algo[i])) {
			ctx->shash_tfm[i] = stfm;
			break;
		}
	}

	This for loop can be replaced with:

	ctx->shash_tfm[sha_req_shash->algo] = stfm;

        By using this we're reducing 5n iterations for n number of
	sha operation.

Bug 200484323
Bug 200485303

Change-Id: Ib4e0e9805502a18da3c365ef2d1f5231035efefd
Signed-off-by: Kartik <kkartik@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2012022
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>