include: increase segement size to 16384
The default size of available to SW-IOMMU is 64MB.
But the maximum allocatable size is IO_TLB_SEGSIZE << IO_TLB_SHIFT,
which is currently 1024 and 11 respectively. Hence, the maximum size
mappable is 1024 << 11 which is nothing but 2 MB.

This is not suffient for display where Linux is running as a guest.
Hence increasing to 16384 which amount to maximum allocatable memory
of 64MB(4096 << 11).

Bug 200490458

Change-Id: If036626225285750f59233491d69430712594c69
Signed-off-by: Ajay Nandakumar <anandakumarm@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2013702
(cherry picked from commit 088717916f581918c7691bf2646ef391082a378f)
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2400884
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit