tegra: camera: rtcpu: use new HSP protocol
Use the new HSP protocol for camera VM if there is a device tree node
nvidia,tegra-camrtc-hsp-vm under the rtcpu node.

Fall back to the ordinary cmd-pair/ivc-pair protocol if no new node is
found or it is disabled.

Clarify the CAMRTC state transitions:
- Handle the nCPUHALT/FWLOADDONE by reset functions
- Wait for CPU idle only when powering off
- Boot sync IVC channels and tracer only with online/offline transition

The CAMRTC firmware stays in sync during runtime suspend. The state is
flushed only when CAMRTC goes offline.  The offline transition happens
when the SoC/VM is suspended or the kernel driver is removed. After
the offline transition, all the shared memory gets reclaimed.

Use the BYE message or PM_COMMAND with non-zero parameter to signal
the SoC/VM suspend to the firmware.

Jira CAMASIL-716
Jira CAMASIL-698

Change-Id: I9a6f1936ad1a2c6e8e94e2af154909daede799c8
Signed-off-by: Pekka Pessi <ppessi@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2018556
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mika Liljeberg <mliljeberg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>