gpu: nvgpu: rfr: print file statistics
Add a listing of changed files after the commit messages similar to how
git show --stat would put it. This makes it trivial to see where the
changes happened if the commit description doesn't state it.

Example from this patch:

  scripts/rfr | 23  (+20  -3  )

The numbers are padded to a width of three so they line up nicely.

Change-Id: Ie62ea6c7d1d3b21446c7447f362d148be3032076
Signed-off-by: Konsta Holtta <kholtta@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2020196
GVS: Gerrit_Virtual_Submit
Reviewed-by: Philip Elcan <pelcan@nvidia.com>
Reviewed-by: Vijayakumar Subbu <vsubbu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>