dts: t194: add dma-mask and iommu-resv-region prop
Add the resv_regions to eqos because it is
added to the domain with audio which requires
a small IOVA range

Add dma-mask property to eqos node. The region from
0x5E000000 to 0x60000000 is required by ADSP for logger
and shared memory. If EQOS driver boots first, this
region becomes unavailable for ADSP driver resulting
in ADSP boot failure. Adding dma-mask property
ensures that the iova range does not overlap with the
region required by ADSP.

Bug 200458098

Change-Id: I76ff86e0c72ee513c97c27769bf178eedcafaace
Reviewed-on: https://git-master.nvidia.com/r/2029058
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Tested-by: Hariharan Sivaraman <hariharans@nvidia.com>
Reviewed-by: Srinivas Ramachandran <srinivasra@nvidia.com>
Reviewed-by: Nitin Pai <npai@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>