gpu: nvgpu: Remove VFE_INVALIDATE RPC for PS3.5
Changes:
In PSTATE3.5, VFE_INVALIDATE is triggered within the PMU
whenever the PMU processes the event. This removes the need
for the driver to explicitly trigger VFE_INVALIDATE.
So, removing VFE_INVALIDATE RPC in pmu_set_boot_clk_runcb_fn
which will be called when we receive a perf event.

Bug 200493291

Change-Id: Ied3e1dd49d148703eb3a067351245a06d0034dcc
Signed-off-by: rmylavarapu <rmylavarapu@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2034476
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svc-mobile-misra <svc-mobile-misra@nvidia.com>
Reviewed-by: svc-misra-checker <svc-misra-checker@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mahantesh Kumbar <mkumbar@nvidia.com>
Reviewed-by: Alex Waterman <alexw@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>