eqos: move PTP register to open
Issue: ptp4l applicaiton is relaying on port /dev/ptp0
not on eth0 interface. Since we are registering the ptp
interface during the probe,/dev/ptp0 will be available
once the probe is done and hence we are seeing a crash
when accessing the PTP API's even though the interface
is down.

Fix: move ptp registration to open instead of probe

Bug 2522054

Change-Id: Iff1101f64a954ccb229a09a6d33b52958f083584
Signed-off-by: Bhadram Varka <vbhadram@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2136293
Reviewed-by: Narayan Reddy <narayanr@nvidia.com>
Tested-by: Narayan Reddy <narayanr@nvidia.com>
Reviewed-by: Rakesh Goyal <rgoyal@nvidia.com>
Reviewed-by: Phoenix Jung <pjung@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Srinivas Ramachandran <srinivasra@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>