net: can: mttcan: use TSC as base time for timestamping
Currently we use PTP as base time for CAN Rx timestamping as CAN
HW counter is only 16 bit wide that can overflow in a second.
For Xavier we are moving to TSC as base time. This is requested
by the stakeholders.
For Parker we will continue to use PTP as base time.

JIRA ESLC-3710

Change-Id: I1caa3d25538806d4a2a54b689b139273ae1c6601
Signed-off-by: Abhijit <abhijit@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2035752
Reviewed-by: Vipin Kumar <vipink@nvidia.com>
Reviewed-by: David Tao <tehyut@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sandeep Trasi <strasi@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>