video: tegra: pva: Deprecate dma_alloc_at_attrs
The dma_alloc_at_attrs is implemented with a hack in the iommu core
code, which is unlikely able to get upstream. So this patch removes
its callers in the pva code and adds an alternative helper function
to do something similar.

The dma_alloc_at_attrs was used to allocate a DMA buffer and to map
it to a specified iova address. And this can be done by remapping a
normal DMA buffer to the iova address.

Bug 200444660

Change-Id: Id73b0a1027dbd475d472ac5983a549df2344bb61
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2083357
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>