gpu: nvgpu: move gk20a_fifo_pbdma_fault_rc to common.rc unit
gk20a_fifo_pbdma_fault_rc is moved to common.rc unit and renamed to
nvgpu_rc_pbdma_fault.

The function is modified such that when the pbdma id is a channel,
recovery is issued only when the channel is part of a valid tsg.

Jira NVGPU-2950

Change-Id: I5e975cf79810479f83ffd50581c214a64d1619a6
Signed-off-by: Debarshi Dutta <ddutta@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2083749
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svc-mobile-misra <svc-mobile-misra@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vijayakumar Subbu <vsubbu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>