video: tegra: pva: Remove iova lucky check
The logic of the pva_dma_alloc_and_map_at() function is to firstly
reserve the target iova and secondly call dma_alloc_attrs() that'd
dynamically allocate another iova.

Since the target iova is reserved, the dynamically allocated iova
is not possible to be identical. And this lucky check lacks of an
exit routine either.

So this patch just drops it.

Bug 200444660

Change-Id: I38b450b1c4faf17045f55edba74ec659556da5a7
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2096706
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>